requirements.txt
Django==5.0.2
djangorestframework==3.14.0
Markdown==3.5.2
django-filter==23.5
psycopg2-binary==2.9.9
gunicorn==21.2.0